backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
httpx==0.27.2
python-dotenv==1.0.1
google-cloud-aiplatform==1.71.1
pydantic==2.9.2
